Enumerations

The enumerations of the C API are available using scoped names, e.g., LLVM.Linkage.Internal for LLVM.API.LLVMInternalLinkage. The values that are available depend on the version of LLVM, so they may differ from the ones listed here.

LLVM.Linkage — Module

LLVM.Linkage

Values of the LLVM.API.LLVMLinkage enum, for the linkage of global values, e.g., LLVM.Linkage.External. Their type is LLVM.Linkage.T. The available values depend on the version of LLVM:

Name	Value	C API

External	0	LLVMExternalLinkage

AvailableExternally	1	LLVMAvailableExternallyLinkage

LinkOnceAny	2	LLVMLinkOnceAnyLinkage

LinkOnceODR	3	LLVMLinkOnceODRLinkage

LinkOnceODRAutoHide	4	LLVMLinkOnceODRAutoHideLinkage

WeakAny	5	LLVMWeakAnyLinkage

WeakODR	6	LLVMWeakODRLinkage

Appending	7	LLVMAppendingLinkage

Internal	8	LLVMInternalLinkage

Private	9	LLVMPrivateLinkage

DLLImport	10	LLVMDLLImportLinkage

DLLExport	11	LLVMDLLExportLinkage

ExternalWeak	12	LLVMExternalWeakLinkage

Ghost	13	LLVMGhostLinkage

Common	14	LLVMCommonLinkage

LinkerPrivate	15	LLVMLinkerPrivateLinkage

LinkerPrivateWeak	16	LLVMLinkerPrivateWeakLinkage

LLVM.Visibility — Module

LLVM.Visibility

Values of the LLVM.API.LLVMVisibility enum, for the visibility of global values, e.g., LLVM.Visibility.Default. Their type is LLVM.Visibility.T. The available values depend on the version of LLVM:

Name	Value	C API

Default	0	LLVMDefaultVisibility

Hidden	1	LLVMHiddenVisibility

Protected	2	LLVMProtectedVisibility

LLVM.DLLStorageClass — Module

LLVM.DLLStorageClass

Values of the LLVM.API.LLVMDLLStorageClass enum, for the DLL storage class of global values, e.g., LLVM.DLLStorageClass.Default. Their type is LLVM.DLLStorageClass.T. The available values depend on the version of LLVM:

Name	Value	C API

Default	0	LLVMDefaultStorageClass

Import	1	LLVMDLLImportStorageClass

Export	2	LLVMDLLExportStorageClass

LLVM.UnnamedAddr — Module

LLVM.UnnamedAddr

Values of the LLVM.API.LLVMUnnamedAddr enum, for whether the address of a global value is significant, e.g., LLVM.UnnamedAddr.No. Their type is LLVM.UnnamedAddr.T. The available values depend on the version of LLVM:

Name	Value	C API

No	0	LLVMNoUnnamedAddr

Local	1	LLVMLocalUnnamedAddr

Global	2	LLVMGlobalUnnamedAddr

LLVM.ThreadLocalMode — Module

LLVM.ThreadLocalMode

Values of the LLVM.API.LLVMThreadLocalMode enum, for the thread-local storage model of global variables, e.g., LLVM.ThreadLocalMode.NotThreadLocal. Their type is LLVM.ThreadLocalMode.T. The available values depend on the version of LLVM:

Name	Value	C API

NotThreadLocal	0	LLVMNotThreadLocal

GeneralDynamic	1	LLVMGeneralDynamicTLSModel

LocalDynamic	2	LLVMLocalDynamicTLSModel

InitialExec	3	LLVMInitialExecTLSModel

LocalExec	4	LLVMLocalExecTLSModel

LLVM.CallConv — Module

LLVM.CallConv

Values of the LLVM.API.LLVMCallConv enum, for calling conventions, e.g., LLVM.CallConv.C. Their type is LLVM.CallConv.T. The available values depend on the version of LLVM:

Name	Value	C API

C	0	LLVMCCallConv

Fast	8	LLVMFastCallConv

Cold	9	LLVMColdCallConv

GHC	10	LLVMGHCCallConv

HiPE	11	LLVMHiPECallConv

AnyReg	13	LLVMAnyRegCallConv

PreserveMost	14	LLVMPreserveMostCallConv

PreserveAll	15	LLVMPreserveAllCallConv

Swift	16	LLVMSwiftCallConv

CXXFASTTLS	17	LLVMCXXFASTTLSCallConv

X86Stdcall	64	LLVMX86StdcallCallConv

X86Fastcall	65	LLVMX86FastcallCallConv

ARMAPCS	66	LLVMARMAPCSCallConv

ARMAAPCS	67	LLVMARMAAPCSCallConv

ARMAAPCSVFP	68	LLVMARMAAPCSVFPCallConv

MSP430INTR	69	LLVMMSP430INTRCallConv

X86ThisCall	70	LLVMX86ThisCallCallConv

PTXKernel	71	LLVMPTXKernelCallConv

PTXDevice	72	LLVMPTXDeviceCallConv

SPIRFUNC	75	LLVMSPIRFUNCCallConv

SPIRKERNEL	76	LLVMSPIRKERNELCallConv

IntelOCLBI	77	LLVMIntelOCLBICallConv

X8664SysV	78	LLVMX8664SysVCallConv

Win64	79	LLVMWin64CallConv

X86VectorCall	80	LLVMX86VectorCallCallConv

HHVM	81	LLVMHHVMCallConv

HHVMC	82	LLVMHHVMCCallConv

X86INTR	83	LLVMX86INTRCallConv

AVRINTR	84	LLVMAVRINTRCallConv

AVRSIGNAL	85	LLVMAVRSIGNALCallConv

AVRBUILTIN	86	LLVMAVRBUILTINCallConv

AMDGPUVS	87	LLVMAMDGPUVSCallConv

AMDGPUGS	88	LLVMAMDGPUGSCallConv

AMDGPUPS	89	LLVMAMDGPUPSCallConv

AMDGPUCS	90	LLVMAMDGPUCSCallConv

AMDGPUKERNEL	91	LLVMAMDGPUKERNELCallConv

X86RegCall	92	LLVMX86RegCallCallConv

AMDGPUHS	93	LLVMAMDGPUHSCallConv

MSP430BUILTIN	94	LLVMMSP430BUILTINCallConv

AMDGPULS	95	LLVMAMDGPULSCallConv

AMDGPUES	96	LLVMAMDGPUESCallConv

LLVM.IntPredicate — Module

LLVM.IntPredicate

Values of the LLVM.API.LLVMIntPredicate enum, for the predicates of integer comparisons, e.g., LLVM.IntPredicate.EQ. Their type is LLVM.IntPredicate.T. The available values depend on the version of LLVM:

Name	Value	C API

EQ	32	LLVMIntEQ

NE	33	LLVMIntNE

UGT	34	LLVMIntUGT

UGE	35	LLVMIntUGE

ULT	36	LLVMIntULT

ULE	37	LLVMIntULE

SGT	38	LLVMIntSGT

SGE	39	LLVMIntSGE

SLT	40	LLVMIntSLT

SLE	41	LLVMIntSLE

LLVM.RealPredicate — Module

LLVM.RealPredicate

Values of the LLVM.API.LLVMRealPredicate enum, for the predicates of floating-point comparisons, e.g., LLVM.RealPredicate.False. Their type is LLVM.RealPredicate.T. The available values depend on the version of LLVM:

Name	Value	C API

False	0	LLVMRealPredicateFalse

OEQ	1	LLVMRealOEQ

OGT	2	LLVMRealOGT

OGE	3	LLVMRealOGE

OLT	4	LLVMRealOLT

OLE	5	LLVMRealOLE

ONE	6	LLVMRealONE

ORD	7	LLVMRealORD

UNO	8	LLVMRealUNO

UEQ	9	LLVMRealUEQ

UGT	10	LLVMRealUGT

UGE	11	LLVMRealUGE

ULT	12	LLVMRealULT

ULE	13	LLVMRealULE

UNE	14	LLVMRealUNE

True	15	LLVMRealPredicateTrue

LLVM.Opcode — Module

LLVM.Opcode

Values of the LLVM.API.LLVMOpcode enum, for the opcodes of instructions and constant expressions, e.g., LLVM.Opcode.Ret. Their type is LLVM.Opcode.T. The available values depend on the version of LLVM:

Name	Value	C API

Ret	1	LLVMRet

Br	2	LLVMBr

Switch	3	LLVMSwitch

IndirectBr	4	LLVMIndirectBr

Invoke	5	LLVMInvoke

Unreachable	7	LLVMUnreachable

Add	8	LLVMAdd

FAdd	9	LLVMFAdd

Sub	10	LLVMSub

FSub	11	LLVMFSub

Mul	12	LLVMMul

FMul	13	LLVMFMul

UDiv	14	LLVMUDiv

SDiv	15	LLVMSDiv

FDiv	16	LLVMFDiv

URem	17	LLVMURem

SRem	18	LLVMSRem

FRem	19	LLVMFRem

Shl	20	LLVMShl

LShr	21	LLVMLShr

AShr	22	LLVMAShr

And	23	LLVMAnd

Or	24	LLVMOr

Xor	25	LLVMXor

Alloca	26	LLVMAlloca

Load	27	LLVMLoad

Store	28	LLVMStore

GetElementPtr	29	LLVMGetElementPtr

Trunc	30	LLVMTrunc

ZExt	31	LLVMZExt

SExt	32	LLVMSExt

FPToUI	33	LLVMFPToUI

FPToSI	34	LLVMFPToSI

UIToFP	35	LLVMUIToFP

SIToFP	36	LLVMSIToFP

FPTrunc	37	LLVMFPTrunc

FPExt	38	LLVMFPExt

PtrToInt	39	LLVMPtrToInt

IntToPtr	40	LLVMIntToPtr

BitCast	41	LLVMBitCast

ICmp	42	LLVMICmp

FCmp	43	LLVMFCmp

PHI	44	LLVMPHI

Call	45	LLVMCall

Select	46	LLVMSelect

UserOp1	47	LLVMUserOp1

UserOp2	48	LLVMUserOp2

VAArg	49	LLVMVAArg

ExtractElement	50	LLVMExtractElement

InsertElement	51	LLVMInsertElement

ShuffleVector	52	LLVMShuffleVector

ExtractValue	53	LLVMExtractValue

InsertValue	54	LLVMInsertValue

Fence	55	LLVMFence

AtomicCmpXchg	56	LLVMAtomicCmpXchg

AtomicRMW	57	LLVMAtomicRMW

Resume	58	LLVMResume

LandingPad	59	LLVMLandingPad

AddrSpaceCast	60	LLVMAddrSpaceCast

CleanupRet	61	LLVMCleanupRet

CatchRet	62	LLVMCatchRet

CatchPad	63	LLVMCatchPad

CleanupPad	64	LLVMCleanupPad

CatchSwitch	65	LLVMCatchSwitch

FNeg	66	LLVMFNeg

CallBr	67	LLVMCallBr

Freeze	68	LLVMFreeze

LLVM.TypeKind — Module

LLVM.TypeKind

Values of the LLVM.API.LLVMTypeKind enum, for the kinds of types, e.g., LLVM.TypeKind.Void. Their type is LLVM.TypeKind.T. The available values depend on the version of LLVM:

Name	Value	C API

Void	0	LLVMVoidTypeKind

Half	1	LLVMHalfTypeKind

Float	2	LLVMFloatTypeKind

Double	3	LLVMDoubleTypeKind

X86_FP80	4	LLVMX86_FP80TypeKind

FP128	5	LLVMFP128TypeKind

PPC_FP128	6	LLVMPPC_FP128TypeKind

Label	7	LLVMLabelTypeKind

Integer	8	LLVMIntegerTypeKind

Function	9	LLVMFunctionTypeKind

Struct	10	LLVMStructTypeKind

Array	11	LLVMArrayTypeKind

Pointer	12	LLVMPointerTypeKind

Vector	13	LLVMVectorTypeKind

Metadata	14	LLVMMetadataTypeKind

X86_MMX	15	LLVMX86_MMXTypeKind

Token	16	LLVMTokenTypeKind

ScalableVector	17	LLVMScalableVectorTypeKind

BFloat	18	LLVMBFloatTypeKind

X86_AMX	19	LLVMX86_AMXTypeKind

TargetExt	20	LLVMTargetExtTypeKind

LLVM.ValueKind — Module

LLVM.ValueKind

Values of the LLVM.API.LLVMValueKind enum, for the kinds of values, e.g., LLVM.ValueKind.Argument. Their type is LLVM.ValueKind.T. The available values depend on the version of LLVM:

Name	Value	C API

Argument	0	LLVMArgumentValueKind

BasicBlock	1	LLVMBasicBlockValueKind

MemoryUse	2	LLVMMemoryUseValueKind

MemoryDef	3	LLVMMemoryDefValueKind

MemoryPhi	4	LLVMMemoryPhiValueKind

Function	5	LLVMFunctionValueKind

GlobalAlias	6	LLVMGlobalAliasValueKind

GlobalIFunc	7	LLVMGlobalIFuncValueKind

GlobalVariable	8	LLVMGlobalVariableValueKind

BlockAddress	9	LLVMBlockAddressValueKind

ConstantExpr	10	LLVMConstantExprValueKind

ConstantArray	11	LLVMConstantArrayValueKind

ConstantStruct	12	LLVMConstantStructValueKind

ConstantVector	13	LLVMConstantVectorValueKind

UndefValue	14	LLVMUndefValueValueKind

ConstantAggregateZero	15	LLVMConstantAggregateZeroValueKind

ConstantDataArray	16	LLVMConstantDataArrayValueKind

ConstantDataVector	17	LLVMConstantDataVectorValueKind

ConstantInt	18	LLVMConstantIntValueKind

ConstantFP	19	LLVMConstantFPValueKind

ConstantPointerNull	20	LLVMConstantPointerNullValueKind

ConstantTokenNone	21	LLVMConstantTokenNoneValueKind

MetadataAsValue	22	LLVMMetadataAsValueValueKind

InlineAsm	23	LLVMInlineAsmValueKind

Instruction	24	LLVMInstructionValueKind

PoisonValue	25	LLVMPoisonValueValueKind

ConstantTargetNone	26	LLVMConstantTargetNoneValueKind

LLVM.AtomicOrdering — Module

LLVM.AtomicOrdering

Values of the LLVM.API.LLVMAtomicOrdering enum, for the orderings of atomic operations, e.g., LLVM.AtomicOrdering.NotAtomic. Their type is LLVM.AtomicOrdering.T. The available values depend on the version of LLVM:

Name	Value	C API

NotAtomic	0	LLVMAtomicOrderingNotAtomic

Unordered	1	LLVMAtomicOrderingUnordered

Monotonic	2	LLVMAtomicOrderingMonotonic

Acquire	4	LLVMAtomicOrderingAcquire

Release	5	LLVMAtomicOrderingRelease

AcquireRelease	6	LLVMAtomicOrderingAcquireRelease

SequentiallyConsistent	7	LLVMAtomicOrderingSequentiallyConsistent

LLVM.AtomicRMWBinOp — Module

LLVM.AtomicRMWBinOp

Values of the LLVM.API.LLVMAtomicRMWBinOp enum, for the operations of atomicrmw instructions, e.g., LLVM.AtomicRMWBinOp.Xchg. Their type is LLVM.AtomicRMWBinOp.T. The available values depend on the version of LLVM:

Name	Value	C API

Xchg	0	LLVMAtomicRMWBinOpXchg

Add	1	LLVMAtomicRMWBinOpAdd

Sub	2	LLVMAtomicRMWBinOpSub

And	3	LLVMAtomicRMWBinOpAnd

Nand	4	LLVMAtomicRMWBinOpNand

Or	5	LLVMAtomicRMWBinOpOr

Xor	6	LLVMAtomicRMWBinOpXor

Max	7	LLVMAtomicRMWBinOpMax

Min	8	LLVMAtomicRMWBinOpMin

UMax	9	LLVMAtomicRMWBinOpUMax

UMin	10	LLVMAtomicRMWBinOpUMin

FAdd	11	LLVMAtomicRMWBinOpFAdd

FSub	12	LLVMAtomicRMWBinOpFSub

FMax	13	LLVMAtomicRMWBinOpFMax

FMin	14	LLVMAtomicRMWBinOpFMin

UIncWrap	15	LLVMAtomicRMWBinOpUIncWrap

UDecWrap	16	LLVMAtomicRMWBinOpUDecWrap

USubCond	17	LLVMAtomicRMWBinOpUSubCond

USubSat	18	LLVMAtomicRMWBinOpUSubSat

FMaximum	19	LLVMAtomicRMWBinOpFMaximum

FMinimum	20	LLVMAtomicRMWBinOpFMinimum

FMaximumNum	21	LLVMAtomicRMWBinOpFMaximumNum

FMinimumNum	22	LLVMAtomicRMWBinOpFMinimumNum

LLVM.TailCallKind — Module

LLVM.TailCallKind

Values of the LLVM.API.LLVMTailCallKind enum, for the tail call markers of calls, e.g., LLVM.TailCallKind.None. Their type is LLVM.TailCallKind.T. The available values depend on the version of LLVM:

Name	Value	C API

None	0	LLVMTailCallKindNone

Tail	1	LLVMTailCallKindTail

MustTail	2	LLVMTailCallKindMustTail

NoTail	3	LLVMTailCallKindNoTail

LLVM.InlineAsmDialect — Module

LLVM.InlineAsmDialect

Values of the LLVM.API.LLVMInlineAsmDialect enum, for the dialects of inline assembly, e.g., LLVM.InlineAsmDialect.ATT. Their type is LLVM.InlineAsmDialect.T. The available values depend on the version of LLVM:

Name	Value	C API

ATT	0	LLVMInlineAsmDialectATT

Intel	1	LLVMInlineAsmDialectIntel

LLVM.ModuleFlagBehavior — Module

LLVM.ModuleFlagBehavior

Values of the LLVM.API.LLVMModuleFlagBehavior enum, for how module flags are merged when linking, e.g., LLVM.ModuleFlagBehavior.Error. Their type is LLVM.ModuleFlagBehavior.T. The available values depend on the version of LLVM:

Name	Value	C API

Error	0	LLVMModuleFlagBehaviorError

Warning	1	LLVMModuleFlagBehaviorWarning

Require	2	LLVMModuleFlagBehaviorRequire

Override	3	LLVMModuleFlagBehaviorOverride

Append	4	LLVMModuleFlagBehaviorAppend

AppendUnique	5	LLVMModuleFlagBehaviorAppendUnique

LLVM.CloneFunctionChangeType — Module

LLVM.CloneFunctionChangeType

Values of the LLVM.API.LLVMCloneFunctionChangeType enum, for the kind of changes when cloning functions, e.g., LLVM.CloneFunctionChangeType.LocalChangesOnly. Their type is LLVM.CloneFunctionChangeType.T. The available values depend on the version of LLVM:

Name	Value	C API

LocalChangesOnly	0	LLVMCloneFunctionChangeTypeLocalChangesOnly

GlobalChanges	1	LLVMCloneFunctionChangeTypeGlobalChanges

DifferentModule	2	LLVMCloneFunctionChangeTypeDifferentModule

ClonedModule	3	LLVMCloneFunctionChangeTypeClonedModule

LLVM.DWARFSourceLanguage — Module

LLVM.DWARFSourceLanguage

Values of the LLVM.API.LLVMDWARFSourceLanguage enum, for the source languages of compile units, e.g., LLVM.DWARFSourceLanguage.C89. Their type is LLVM.DWARFSourceLanguage.T. The available values depend on the version of LLVM:

Name	Value	C API

C89	0	LLVMDWARFSourceLanguageC89

C	1	LLVMDWARFSourceLanguageC

Ada83	2	LLVMDWARFSourceLanguageAda83

C_plus_plus	3	LLVMDWARFSourceLanguageC_plus_plus

Cobol74	4	LLVMDWARFSourceLanguageCobol74

Cobol85	5	LLVMDWARFSourceLanguageCobol85

Fortran77	6	LLVMDWARFSourceLanguageFortran77

Fortran90	7	LLVMDWARFSourceLanguageFortran90

Pascal83	8	LLVMDWARFSourceLanguagePascal83

Modula2	9	LLVMDWARFSourceLanguageModula2

Java	10	LLVMDWARFSourceLanguageJava

C99	11	LLVMDWARFSourceLanguageC99

Ada95	12	LLVMDWARFSourceLanguageAda95

Fortran95	13	LLVMDWARFSourceLanguageFortran95

PLI	14	LLVMDWARFSourceLanguagePLI

ObjC	15	LLVMDWARFSourceLanguageObjC

ObjC_plus_plus	16	LLVMDWARFSourceLanguageObjC_plus_plus

UPC	17	LLVMDWARFSourceLanguageUPC

D	18	LLVMDWARFSourceLanguageD

Python	19	LLVMDWARFSourceLanguagePython

OpenCL	20	LLVMDWARFSourceLanguageOpenCL

Go	21	LLVMDWARFSourceLanguageGo

Modula3	22	LLVMDWARFSourceLanguageModula3

Haskell	23	LLVMDWARFSourceLanguageHaskell

C_plus_plus_03	24	LLVMDWARFSourceLanguageC_plus_plus_03

C_plus_plus_11	25	LLVMDWARFSourceLanguageC_plus_plus_11

OCaml	26	LLVMDWARFSourceLanguageOCaml

Rust	27	LLVMDWARFSourceLanguageRust

C11	28	LLVMDWARFSourceLanguageC11

Swift	29	LLVMDWARFSourceLanguageSwift

Julia	30	LLVMDWARFSourceLanguageJulia

Dylan	31	LLVMDWARFSourceLanguageDylan

C_plus_plus_14	32	LLVMDWARFSourceLanguageC_plus_plus_14

Fortran03	33	LLVMDWARFSourceLanguageFortran03

Fortran08	34	LLVMDWARFSourceLanguageFortran08

RenderScript	35	LLVMDWARFSourceLanguageRenderScript

BLISS	36	LLVMDWARFSourceLanguageBLISS

Kotlin	37	LLVMDWARFSourceLanguageKotlin

Zig	38	LLVMDWARFSourceLanguageZig

Crystal	39	LLVMDWARFSourceLanguageCrystal

C_plus_plus_17	40	LLVMDWARFSourceLanguageC_plus_plus_17

C_plus_plus_20	41	LLVMDWARFSourceLanguageC_plus_plus_20

C17	42	LLVMDWARFSourceLanguageC17

Fortran18	43	LLVMDWARFSourceLanguageFortran18

Ada2005	44	LLVMDWARFSourceLanguageAda2005

Ada2012	45	LLVMDWARFSourceLanguageAda2012

Mojo	46	LLVMDWARFSourceLanguageMojo

Mips_Assembler	47	LLVMDWARFSourceLanguageMips_Assembler

GOOGLE_RenderScript	48	LLVMDWARFSourceLanguageGOOGLE_RenderScript

BORLAND_Delphi	49	LLVMDWARFSourceLanguageBORLAND_Delphi

LLVM.DWARFEmissionKind — Module

LLVM.DWARFEmissionKind

Values of the LLVM.API.LLVMDWARFEmissionKind enum, for the amount of debug info to emit for compile units, e.g., LLVM.DWARFEmissionKind.None. Their type is LLVM.DWARFEmissionKind.T. The available values depend on the version of LLVM:

Name	Value	C API

None	0	LLVMDWARFEmissionNone

Full	1	LLVMDWARFEmissionFull

LineTablesOnly	2	LLVMDWARFEmissionLineTablesOnly

LLVM.DebugEmissionKind — Module

LLVM.DebugEmissionKind

Values of the LLVM.API.LLVMDebugEmissionKind enum, for the amount of debug info that Julia's code generator emits (debug_info_kind in its CodegenParams), e.g., LLVM.DebugEmissionKind.NoDebug. Their type is LLVM.DebugEmissionKind.T. The available values depend on the version of LLVM:

Name	Value	C API

NoDebug	0	LLVMDebugEmissionKindNoDebug

FullDebug	1	LLVMDebugEmissionKindFullDebug

LineTablesOnly	2	LLVMDebugEmissionKindLineTablesOnly

DebugDirectivesOnly	3	LLVMDebugEmissionKindDebugDirectivesOnly

LLVM.CodeGenOptLevel — Module

LLVM.CodeGenOptLevel

Values of the LLVM.API.LLVMCodeGenOptLevel enum, for the optimization levels of code generation, e.g., LLVM.CodeGenOptLevel.None. Their type is LLVM.CodeGenOptLevel.T. The available values depend on the version of LLVM:

Name	Value	C API

None	0	LLVMCodeGenLevelNone

Less	1	LLVMCodeGenLevelLess

Default	2	LLVMCodeGenLevelDefault

Aggressive	3	LLVMCodeGenLevelAggressive

LLVM.CodeGenFileType — Module

LLVM.CodeGenFileType

Values of the LLVM.API.LLVMCodeGenFileType enum, for the kinds of files that code generation emits, e.g., LLVM.CodeGenFileType.Assembly. Their type is LLVM.CodeGenFileType.T. The available values depend on the version of LLVM:

Name	Value	C API

Assembly	0	LLVMAssemblyFile

Object	1	LLVMObjectFile

LLVM.RelocMode — Module

LLVM.RelocMode

Values of the LLVM.API.LLVMRelocMode enum, for the relocation models of target machines, e.g., LLVM.RelocMode.Default. Their type is LLVM.RelocMode.T. The available values depend on the version of LLVM:

Name	Value	C API

Default	0	LLVMRelocDefault

Static	1	LLVMRelocStatic

PIC	2	LLVMRelocPIC

DynamicNoPic	3	LLVMRelocDynamicNoPic

ROPI	4	LLVMRelocROPI

RWPI	5	LLVMRelocRWPI

ROPI_RWPI	6	LLVMRelocROPI_RWPI

LLVM.CodeModel — Module

LLVM.CodeModel

Values of the LLVM.API.LLVMCodeModel enum, for the code models of target machines, e.g., LLVM.CodeModel.Default. Their type is LLVM.CodeModel.T. The available values depend on the version of LLVM:

Name	Value	C API

Default	0	LLVMCodeModelDefault

JITDefault	1	LLVMCodeModelJITDefault

Tiny	2	LLVMCodeModelTiny

Small	3	LLVMCodeModelSmall

Kernel	4	LLVMCodeModelKernel

Medium	5	LLVMCodeModelMedium

Large	6	LLVMCodeModelLarge

LLVM.ByteOrdering — Module

LLVM.ByteOrdering

Values of the LLVM.API.LLVMByteOrdering enum, for the byte orders of data layouts, e.g., LLVM.ByteOrdering.Big. Their type is LLVM.ByteOrdering.T. The available values depend on the version of LLVM:

Name	Value	C API

Big	0	LLVMBigEndian

Little	1	LLVMLittleEndian

LLVM.LookupKind — Module

LLVM.LookupKind

Values of the LLVM.API.LLVMOrcLookupKind enum, for the kinds of ORC symbol lookups, e.g., LLVM.LookupKind.Static. Their type is LLVM.LookupKind.T. The available values depend on the version of LLVM:

Name	Value	C API

Static	0	LLVMOrcLookupKindStatic

DLSym	1	LLVMOrcLookupKindDLSym

LLVM.JITDylibLookupFlags — Module

LLVM.JITDylibLookupFlags

Values of the LLVM.API.LLVMOrcJITDylibLookupFlags enum, for which symbols of a JITDylib an ORC lookup matches, e.g., LLVM.JITDylibLookupFlags.MatchExportedSymbolsOnly. Their type is LLVM.JITDylibLookupFlags.T. The available values depend on the version of LLVM:

Name	Value	C API

MatchExportedSymbolsOnly	0	LLVMOrcJITDylibLookupFlagsMatchExportedSymbolsOnly

MatchAllSymbols	1	LLVMOrcJITDylibLookupFlagsMatchAllSymbols

LLVM.SymbolLookupFlags — Module

LLVM.SymbolLookupFlags

Values of the LLVM.API.LLVMOrcSymbolLookupFlags enum, for whether an ORC lookup requires a symbol, e.g., LLVM.SymbolLookupFlags.RequiredSymbol. Their type is LLVM.SymbolLookupFlags.T. The available values depend on the version of LLVM:

Name	Value	C API

RequiredSymbol	0	LLVMOrcSymbolLookupFlagsRequiredSymbol

WeaklyReferencedSymbol	1	LLVMOrcSymbolLookupFlagsWeaklyReferencedSymbol