Instruction types

The types of the instructions that LLVM supports.

LLVM.RetInst — Type

LLVM.RetInst <: Instruction

The type of ret instructions.

LLVM.BrInst — Type

LLVM.BrInst <: Instruction

The type of br instructions.

LLVM.SwitchInst — Type

LLVM.SwitchInst <: Instruction

The type of switch instructions.

LLVM.IndirectBrInst — Type

LLVM.IndirectBrInst <: Instruction

The type of indirectbr instructions.

LLVM.InvokeInst — Type

LLVM.InvokeInst <: Instruction

The type of invoke instructions.

LLVM.UnreachableInst — Type

LLVM.UnreachableInst <: Instruction

The type of unreachable instructions.

LLVM.CallBrInst — Type

LLVM.CallBrInst <: Instruction

The type of callbr instructions.

LLVM.FNegInst — Type

LLVM.FNegInst <: Instruction

The type of fneg instructions.

LLVM.AddInst — Type

LLVM.AddInst <: Instruction

The type of add instructions.

LLVM.FAddInst — Type

LLVM.FAddInst <: Instruction

The type of fadd instructions.

LLVM.SubInst — Type

LLVM.SubInst <: Instruction

The type of sub instructions.

LLVM.FSubInst — Type

LLVM.FSubInst <: Instruction

The type of fsub instructions.

LLVM.MulInst — Type

LLVM.MulInst <: Instruction

The type of mul instructions.

LLVM.FMulInst — Type

LLVM.FMulInst <: Instruction

The type of fmul instructions.

LLVM.UDivInst — Type

LLVM.UDivInst <: Instruction

The type of udiv instructions.

LLVM.SDivInst — Type

LLVM.SDivInst <: Instruction

The type of sdiv instructions.

LLVM.FDivInst — Type

LLVM.FDivInst <: Instruction

The type of fdiv instructions.

LLVM.URemInst — Type

LLVM.URemInst <: Instruction

The type of urem instructions.

LLVM.SRemInst — Type

LLVM.SRemInst <: Instruction

The type of srem instructions.

LLVM.FRemInst — Type

LLVM.FRemInst <: Instruction

The type of frem instructions.

LLVM.ShlInst — Type

LLVM.ShlInst <: Instruction

The type of shl instructions.

LLVM.LShrInst — Type

LLVM.LShrInst <: Instruction

The type of lshr instructions.

LLVM.AShrInst — Type

LLVM.AShrInst <: Instruction

The type of ashr instructions.

LLVM.AndInst — Type

LLVM.AndInst <: Instruction

The type of and instructions.

LLVM.OrInst — Type

LLVM.OrInst <: Instruction

The type of or instructions.

LLVM.XorInst — Type

LLVM.XorInst <: Instruction

The type of xor instructions.

LLVM.AllocaInst — Type

LLVM.AllocaInst <: Instruction

The type of alloca instructions.

LLVM.LoadInst — Type

LLVM.LoadInst <: Instruction

The type of load instructions.

LLVM.StoreInst — Type

LLVM.StoreInst <: Instruction

The type of store instructions.

LLVM.GetElementPtrInst — Type

LLVM.GetElementPtrInst <: Instruction

The type of getelementptr instructions.

LLVM.TruncInst — Type

LLVM.TruncInst <: Instruction

The type of trunc instructions.

LLVM.ZExtInst — Type

LLVM.ZExtInst <: Instruction

The type of zext instructions.

LLVM.SExtInst — Type

LLVM.SExtInst <: Instruction

The type of sext instructions.

LLVM.FPToUIInst — Type

LLVM.FPToUIInst <: Instruction

The type of fptoui instructions.

LLVM.FPToSIInst — Type

LLVM.FPToSIInst <: Instruction

The type of fptosi instructions.

LLVM.UIToFPInst — Type

LLVM.UIToFPInst <: Instruction

The type of uitofp instructions.

LLVM.SIToFPInst — Type

LLVM.SIToFPInst <: Instruction

The type of sitofp instructions.

LLVM.FPTruncInst — Type

LLVM.FPTruncInst <: Instruction

The type of fptrunc instructions.

LLVM.FPExtInst — Type

LLVM.FPExtInst <: Instruction

The type of fpext instructions.

LLVM.PtrToIntInst — Type

LLVM.PtrToIntInst <: Instruction

The type of ptrtoint instructions.

LLVM.IntToPtrInst — Type

LLVM.IntToPtrInst <: Instruction

The type of inttoptr instructions.

LLVM.BitCastInst — Type

LLVM.BitCastInst <: Instruction

The type of bitcast instructions.

LLVM.AddrSpaceCastInst — Type

LLVM.AddrSpaceCastInst <: Instruction

The type of addrspacecast instructions.

LLVM.ICmpInst — Type

LLVM.ICmpInst <: Instruction

The type of icmp instructions.

LLVM.FCmpInst — Type

LLVM.FCmpInst <: Instruction

The type of fcmp instructions.

LLVM.PHIInst — Type

LLVM.PHIInst <: Instruction

The type of phi instructions.

LLVM.CallInst — Type

LLVM.CallInst <: Instruction

The type of call instructions.

LLVM.SelectInst — Type

LLVM.SelectInst <: Instruction

The type of select instructions.

LLVM.UserOp1Inst — Type

LLVM.UserOp1Inst <: Instruction

An instruction that only exists internally, while some of LLVM's passes run.

LLVM.UserOp2Inst — Type

LLVM.UserOp2Inst <: Instruction

An instruction that only exists internally, while some of LLVM's passes run.

LLVM.VAArgInst — Type

LLVM.VAArgInst <: Instruction

The type of va_arg instructions.

LLVM.ExtractElementInst — Type

LLVM.ExtractElementInst <: Instruction

The type of extractelement instructions.

LLVM.InsertElementInst — Type

LLVM.InsertElementInst <: Instruction

The type of insertelement instructions.

LLVM.ShuffleVectorInst — Type

LLVM.ShuffleVectorInst <: Instruction

The type of shufflevector instructions.

LLVM.ExtractValueInst — Type

LLVM.ExtractValueInst <: Instruction

The type of extractvalue instructions.

LLVM.InsertValueInst — Type

LLVM.InsertValueInst <: Instruction

The type of insertvalue instructions.

LLVM.FreezeInst — Type

LLVM.FreezeInst <: Instruction

The type of freeze instructions.

LLVM.FenceInst — Type

LLVM.FenceInst <: Instruction

The type of fence instructions.

LLVM.AtomicCmpXchgInst — Type

LLVM.AtomicCmpXchgInst <: Instruction

The type of cmpxchg instructions.

LLVM.AtomicRMWInst — Type

LLVM.AtomicRMWInst <: Instruction

The type of atomicrmw instructions.

LLVM.ResumeInst — Type

LLVM.ResumeInst <: Instruction

The type of resume instructions.

LLVM.LandingPadInst — Type

LLVM.LandingPadInst <: Instruction

The type of landingpad instructions.

LLVM.CleanupRetInst — Type

LLVM.CleanupRetInst <: Instruction

The type of cleanupret instructions.

LLVM.CatchRetInst — Type

LLVM.CatchRetInst <: Instruction

The type of catchret instructions.

LLVM.CatchPadInst — Type

LLVM.CatchPadInst <: Instruction

The type of catchpad instructions.

LLVM.CleanupPadInst — Type

LLVM.CleanupPadInst <: Instruction

The type of cleanuppad instructions.

LLVM.CatchSwitchInst — Type

LLVM.CatchSwitchInst <: Instruction

The type of catchswitch instructions.