Basic blocks

LLVM.BasicBlock — Type

BasicBlock

A basic block in the IR. A basic block is a sequence of instructions that always ends in a terminator instruction.

LLVM.BasicBlock — Method

BasicBlock(name::String)

Create a new, empty basic block with the given name.

LLVM.BasicBlock — Method

BasicBlock(f::LLVM.Function, name::String)

Create a new, empty basic block with the given name, and insert it at the end of the given function.

LLVM.BasicBlock — Method

BasicBlock(bb::BasicBlock, name::String)

Create a new, empty basic block with the given name, and insert it before the given basic block.

Properties and operations

LLVM.remove! — Method

remove!(bb::BasicBlock)

Remove the given basic block from its parent function, but do not free the object.

LLVM.erase! — Method

erase!(fun::Function, bb::BasicBlock)

Remove the given basic block from its parent function and free the object.

Warning

This function is unsafe because it does not check if the basic block is used elsewhere.

LLVM.parent — Method

parent(bb::BasicBlock) -> LLVM.Function

Get the function that contains the given basic block, or nothing if the block is not part of a function.

LLVM.terminator — Method

terminator(bb::BasicBlock) -> LLVM.Instruction

Get the terminator instruction of the given basic block.

LLVM.name — Method

name(bb::BasicBlock) -> String

Get the name of the given basic block.

LLVM.move_before — Function

move_before(bb::BasicBlock, pos::BasicBlock)

Move the given basic block before the given position.

LLVM.move_after — Function

move_after(bb::BasicBlock, pos::BasicBlock)

Move the given basic block after the given position.

Control flow

LLVM.predecessors — Method

predecessors(bb::BasicBlock)

Get the predecessors of the given basic block.

LLVM.successors — Method

successors(bb::BasicBlock)

Get the successors of the given basic block.

Instructions

LLVM.instructions — Function

instructions(bb::BasicBlock)

Get an iterator over the instructions in the given basic block.

LLVM.previnst — Function

previnst(inst::Instruction)

Get the instruction before the given instruction in the basic block, or nothing if there is none.

LLVM.nextinst — Function

nextinst(inst::Instruction)

Get the instruction after the given instruction in the basic block, or nothing if there is none.