Julia/LLVM interop

Julia/LLVM interop

This section lists functionality for connecting Julia with LLVM.jl, e.g. emitting code for the Julia JIT or creating types that are compatible with Julia's global state.

Base functionality

LLVM.Interop.JuliaContext — Function.

JuliaContext()

Returns the (session-bound) LLVM context used by the Julia compiler.

source

Base.convert — Method.

convert(LLVMType, typ::Type)

Convert a Julia type typ to its LLVM representation. Fails if the type would be boxed.

source

LLVM.Interop.create_function — Function.

create_function(rettyp::LLVMType, argtyp::Vector{LLVMType}, [name::String])

Create an LLVM function, given its return type rettyp and a vector of argument types argtyp. The function is marked for inlining, to be embedded in the caller's body. Returns both the newly created function, and its type.

source

LLVM.Interop.call_function — Function.

call_function(f::LLVM.Function, rettyp::Type, argtyp::Type, args::Expr)

Generate a call to an LLVM function f, given its return type rettyp and a tuple-type for the arguments. The arguments should be passed as an expression yielding a tuple of the argument values (eg. :((1,2))), which will be splatted into the call to the function.

source

Inline assembly

LLVM.Interop.@asmcall — Macro.

@asmcall asm::String [constraints::String] [side_effects::Bool=false]
         rettyp=Nothing argtyp=Tuple{} args...

Call some inline assembly asm, optionally constrained by constraints and denoting other side effects in side_effects, specifying the return type in rettyp and types of arguments as a tuple-type in argtyp.

source