Metadata

LLVM.Metadata — Type

Metadata

Abstract supertype for all metadata types.

LLVM.MDString — Type

MDString

A string metadata node.

LLVM.MDString — Method

MDString(val::String)

Create a new string metadata node from the given Julia string.

Base.convert — Method

convert(String, md::MDString)

Get the string value of the given string metadata node.

LLVM.MDNode — Type

MDNode

Abstract supertype for metadata nodes that can have operands.

See also: MDTuple for a concrete subtype.

LLVM.operands — Method

operands(md::MDNode)

Get the operands of the given metadata node.

LLVM.MDTuple — Type

MDTuple

A tuple metadata node.

LLVM.MDNode — Method

MDNode(vals::Vector) -> MDTuple

Create a new tuple metadata node from the given operands.

Passing nothing as a value will result in a null operand.

Metadata <-> Value

LLVM.ValueAsMetadata — Type

LLVM.ValueAsMetadata

Abstract type for values wrapped as metadata, for use in APIs that expect a LLVM.Metadata.

See also: Metadata(::Value) to convert back to a metadata.

LLVM.Metadata — Method

Metadata(val::Value)

Wrap the given value as metadata, for use in APIs that expect a LLVM.Metadata.

When the value is already metadata wrapped as a value, this will simply return the original metadata.

LLVM.MetadataAsValue — Type

LLVM.MetadataAsValue

Metadata wrapped as a regular value, for use in APIs that expect a LLVM.Value.

See also: Value(::Metadata) to convert back to a value.

LLVM.Value — Method

Value(md::Metadata)

Wrap the given metadata as a value, for use in APIs that expect a LLVM.Value.

When the metadata is already a value wrapped as metadata, this will simply return the original value.

Inspecting and attaching

LLVM.metadata — Method

metadata(inst::Instruction)
metadata(inst::GlobalObject)

Iterate over the metadata of the given instruction or global object.

These iterators are mutable, and implement setindex! and delete! to modify the metadata.

LLVM.NamedMDNode — Type

NamedMDNode

A named metadata node, which is a collection of metadata nodes with a name.

LLVM.metadata — Method

metadata(mod)

Fetch the module-level named metadata. This can be inspected using a Dict-like interface. Mutation is different: There is no setindex! method, as named metadata is append-only. Instead, fetch the named metadata node using getindex, and push! to it.

LLVM.name — Method

name(node::NamedMDNode)

Get the name of the given named metadata node.

LLVM.operands — Method

operands(node::NamedMDNode)

Get the operands of the given named metadata node.

Base.push! — Method

push!(node::NamedMDNode, val::MDNode)

Add a metadata node to the given named metadata node.

Debug information

LLVM.DINode — Type

DINode

a tagged DWARF-like metadata node.

Location information

LLVM.DILocation — Type

DILocation

A location in the source code.

LLVM.line — Method

line(location::DILocation)

Get the line number of this debug location.

LLVM.column — Function

column(location::DILocation)

Get the column number of this debug location.

LLVM.scope — Method

scope(location::DILocation)

Get the local scope associated with this debug location.

LLVM.inlined_at — Function

inlined_at(location::DILocation)

Get the "inline at" location associated with this debug location.

Variables

LLVM.DIVariable — Type

DIVariable

Abstract supertype for all variable-like metadata nodes.

LLVM.DILocalVariable — Type

DILocalVariable <: DIVariable

A local variable in the source code.

LLVM.DIGlobalVariable — Type

DIGlobalVariable <: DIVariable

A global variable in the source code.

LLVM.file — Method

file(var::DIVariable)

Get the file of the given variable.

LLVM.scope — Method

name(var::DIVariable)

Get the name of the given variable.

LLVM.line — Method

line(var::DIVariable)

Get the line number of the given variable.

Scopes

LLVM.DIScope — Type

DIScope

Abstract supertype for lexical scopes and types (which are also declaration contexts).

LLVM.file — Method

file(scope::DIScope)

Get the metadata of the file associated with a given scope.

LLVM.name — Method

name(scope::DIScope)

Get the name of the given scope.

File

LLVM.DIFile — Type

DIFile

A file in the source code.

LLVM.directory — Function

directory(file::DIFile)

Get the directory of a given file.

LLVM.filename — Function

filename(file::DIFile)

Get the filename of the given file.

LLVM.source — Function

source(file::DIFile)

Get the source of the given file, or nothing if the source is not available.

Type

LLVM.DIType — Type

DIType

Abstract supertype for all type-like metadata nodes.

LLVM.name — Method

name(typ::DIType)

Get the name of the given type.

Base.sizeof — Method

sizeof(typ::DIType)

Get the size in bits of the given type.

LLVM.offset — Method

offset(typ::DIType)

Get the offset in bits of the given type.

LLVM.line — Method

line(typ::DIType)

Get the line number of the given type.

LLVM.flags — Method

flags(typ::DIType)

Get the flags of the given type.

Subprogram

LLVM.DISubProgram — Type

DISubProgram

A subprogram in the source code.

LLVM.line — Method

line(subprogram::DISubProgram)

Get the line number of the given subprogram.

Compile Unit

LLVM.DICompileUnit — Type

DICompileUnit

A compilation unit in the source code.

Other

LLVM.DEBUG_METADATA_VERSION — Function

DEBUG_METADATA_VERSION()

The current debug info version number, as supported by LLVM.

LLVM.strip_debuginfo! — Function

strip_debuginfo!(mod::Module)

Strip the debug information from the given module.

LLVM.subprogram — Method

subprogram(func::Function) -> DISubProgram

Get the subprogram of the given function, or nothing if the function has no subprogram.

LLVM.subprogram! — Function

subprogram!(func::Function, sp::DISubProgram)

Set the subprogram of the given function.